x-scraper-scrapling/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
scrapling[fetchers]==0.4.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.0
alembic==1.14.0
httpx==0.27.2
